% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_ppEffect.R
\name{corr_ppEffect}
\alias{corr_ppEffect}
\title{Correct ppEffect}
\usage{
corr_ppEffect(
  object,
  ppDEGs,
  method = "regress.out.ppDEGs",
  pp.Score.TH = "pp.Score Q3"
)
}
\arguments{
\item{object}{Seurat object,have been well pre-propcessed}

\item{ppDEGs}{a vector of protoplasting induced genes}

\item{method}{we provided several methods to correct the ppDEGs,
Options include 'regress.out.ppDEGs', 'remove.ppCells','remove.ppDEGs'}

\item{pp.Score.TH}{You can choose the threshold of pp.Score when try to remove the cells which were
highly stimulated by protoplasting. The default one is}
}
\value{
a new seruat object after ppEffect correction.
}
\description{
Correct ppEffect
}
\examples{
# object_corr <- corr_ppEffect(object = object,ppDEGs =ppDEGs, method = "regress.out.ppDEGs")
}
