% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppDEGs_DB.R
\name{Details}
\alias{Details}
\title{View the details about the ppDEGs_DB}
\usage{
Details(x, ID)
}
\arguments{
\item{x}{a object in ppDEGs_DB class}

\item{ID}{ID of the dataset you interested, E.g "01"}
}
\description{
View the details about the ppDEGs_DB
}
