% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppDEGs_DB.R
\name{Overview}
\alias{Overview}
\title{Print the datasets overview in ppDEGs_DB}
\usage{
Overview(x)
}
\arguments{
\item{x}{an object in ppDEGs_DB class}
}
\description{
Print the datasets overview in ppDEGs_DB
}
