% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppDEGs.R
\name{ppDEGs}
\alias{ppDEGs}
\title{Define a S3 object in class of ppDEGs to store each dataset}
\usage{
ppDEGs(ID, name, species, sample, treatment, ref, genes)
}
\arguments{
\item{ID}{ID number of the dataset}

\item{name}{name of the dataset,E.g "Species_Tissue_Author_Year"}

\item{species}{Species}

\item{sample}{Tissue or Organs of the sample}

\item{treatment}{Time,temperature, and othe protoplasting conditions}

\item{ref}{DOI of the ref.}

\item{genes}{a vector of ppDEGs}
}
\description{
Define a S3 object in class of ppDEGs to store each dataset
}
\examples{
 # A1 <- ppDEGs(ID="01",
 #              name = "ara_root_Wang_2020",
 #              species = "ara",
 #              sample = "root",
 #              treatment ="3H",
 #              ref = "DOI:001",
 #              genes = c("ATgene1","ATgene2","ATgene3"))

}
