% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppDEGs_DB.R
\name{ppDEGs_DB}
\alias{ppDEGs_DB}
\title{Define the class of ppDEGs_DB to store multiply ppDEGs datasets}
\usage{
ppDEGs_DB(ppDEGs)
}
\arguments{
\item{ppDEGs}{a object of ppDEGs}
}
\description{
Define the class of ppDEGs_DB to store multiply ppDEGs datasets
}
\examples{
# A1 <- ppDEGs(ID="01",name = "ara_root_Wang_2020",species = "ara",sample = "root",
# treatment ="3H",ref = "DOI:001",genes = c("ATgene1","ATgene2","ATgene3"))

# A2 <- ppDEGs(ID="02",name = "Os_root_Yao_2022",
# species = "Os",sample = "root",treatment ="1H",ref = "DOI:002",genes = c("Osgene1","Osgene2","Osgene3"))

ppDEGs_DB <- ppDEGs_DB(list("01"=A1,"02"=A2))
}
