% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppDEGs_DB.R
\name{ppDEGsExtra}
\alias{ppDEGsExtra}
\title{Extra ppDEGs}
\usage{
ppDEGsExtra(x, ID)
}
\arguments{
\item{x}{a object in ppDEGs_DB class}

\item{ID}{ID of the dataset you interested, E.g "01"}
}
\description{
Extra ppDEGs
}
